+++ ext/EMHInvExt.jl
@@ -1,0 +1,17 @@
+module EMHInvExt
+
+using EnergyModelsBase
+using EnergyModelsHydrogen
+using EnergyModelsInvestments
+using JuMP
+using TimeStruct
+
+const EMB = EnergyModelsBase
+const EMH = EnergyModelsHydrogen
+const EMI = EnergyModelsInvestments
+const TS = TimeStruct
+
+include("model.jl")
+include("utils.jl")
+
+end